backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weasyprint==67.0
pydantic==2.5.0
python-multipart==0.0.6
psutil==5.9.6
websockets==12.0
pyyaml==6.0.1
